fix: fixing on seller wrapper when there is no seller available refactor: refactoring and adding some custom hooks
